Create deleteListItem
After prompting for confirmation, deletes a list item to the recycle bin, then deletes a folder corresponding to that list item (one folder per item in this application). You can borrow this code to delete a list item, or delete a folder, or both. Deleting files from folders is similar, but not listed here. In each case the /recycle at the end of the REST endpoint causes the deleted object to go to the site's recycle bin.